requirements.txt
fastapi
uvicorn
rank_bm25
pydantic
jinja2
python-multipart
sentence-transformers
faiss-cpu
beautifulsoup4\nrequests
